remove old management commands
